linux-compulab: fix defconfig generation
Before this change the default feature of adding kernel defconfig
fragments in form of .cfg / .scc was broken because do_configure would
call make to generate a .config out of a defconfig and a single fragment
in-tree, completely disregarding all the workd done by
kernel-yocto.bbclass

To fix this we merge the in-tree defconfig with the in-tree fragments
using do_merge_delta_config task (copied from meta-freescale) and save
that as defconfig in ${WORKDIR}. kernel-yocto will then take this file
as a defconfig base and apply any fragments specified in SRC_URI.

This fixes the previously two fragments not being used (
linux-firmware-sdma.cfg, linux-headers.cfg).

Also, don't set LINUX_VERSION_EXTENSION. The in-tree fragment already
sets CONFIG_LOCALVERSION="-iot-gate-imx8m-4.0.0", this would just
overwrite it with "-4.0" which I don't think was desired.